Refactor/a better foramt to handle message
