+++ bns-core/src/lib.rs
@@ -3,9 +3,9 @@
 #![feature(box_syntax)]
 
 pub mod channels;
-pub mod dht;
 pub mod ecc;
 pub mod message;
+pub mod routing;
 pub mod storage;
 pub mod swarm;
 pub mod transports;
