+++ bns-core/src/routing/mod.rs
@@ -1,0 +1,5 @@
+mod did;
+pub use did::Did;
+
+mod chord;
+pub use chord::Chord;
